This is My Website
Title: Logic School of Arts – School of Arts in Z World

Concept:

The project aimed to create a website for a fictitious school situated in an alternate world called "Z World." This world redefined all norms and was discovered by artists in the year 1050.
The aim was to offer visitors an immersive experience that complements the video we did in class. The website has information about the school and its history. It also displays the video depicting the unique artistic approach of art in the school and widely in the world where it is situated. Also, the website allows interested individuals of all ages to sign in to learn more and eventually enroll in the school and enroll in it. The theme of this project is to convey mystery and intrigue visitors because it is a place where nothing is what it seems. To emphasize this a ghost was added to the website.

Process:

The implementation of my idea presented significant challenges. First and foremost, I did not know how to code, and I suddenly was doing HTML, CSS, and JavaScript. I was overwhelmed and scared at the beginning. I feel I had a limited time to explore these languages, but enough to get me started. I spent hours upon hours watching videos reading about these languages, and exploring. I invested countless hours in watching tutorials, reading documentation, and seeking examples online. While I made progress in a short span, web development is a constantly evolving field, and there is always more to learn.

 I think that the website development was both rewarding and frustrating. I had to examine and debug every single line to figure out errors and why things were not working. One extra coma, typo, or space would ruin the whole website and I had to squeeze my eyes to find it. This challenging experience made me realize how much I can do with little knowledge of coding. Who knew I could design a web? Editing and embedding the video was the easy part I would say.
During the implementation, I realized the best way to do the website is to think about each part within it in terms of blocks of a puzzle and my job is to put it together. I was systematic about it. I would do both HTML, CSS, and JavaScript together to make sure I get exactly what I want. If something does not work, I would comment on it dawn and get to it later. Google, YouTube, resources from Brightspace, and w3school were my friends through the whole process.

Reflection:

I am proud of the results. I think I was able to make the website work with no errors and support the video. It is esthetically simple with a clear idea. I think that the website successfully created a sense of intrigue and curiosity, as desired. However, I faced a lot of challenges and moments of frustration. I will say I learned a lot in this journey, especially in a short time.

Web development is an evolving field and consistent learning is important. I am excited to learn more about it because there is always room for improvement. I am sure that most of the things I have done have shortcuts that I still need to discover.